resolves TODO: run dotfile forest less often
dotfile "forest" of ~/.host/ is a "forest" because the dotfiles are
broken up into individual "trees" which each handle a task
({distro,host,deskenv,domain}) by invoking various logic internally.

Most of this logic needs to be run once in a while, not every single
shell invocation. So this TODO is now resolved because we run each
"tree" only if *that* tree hasn't been run before (that is: we're not
naively checking if the _forest_ has been walked, we're checking on each
_tree_ to see if it's been handled). This means I can log into a tty
seat (not invoking DE (desktop environment) specific logic), then login
later to a DE and that DE's logic will run exactly once as intended.